requirements.txt
# ========================================
# FLASK & WEB FRAMEWORK
# ========================================
Flask==3.0.0
Werkzeug==3.0.0

# Environment Variables
python-dotenv==1.0.0

# Security & Rate Limiting
Flask-Limiter==3.5.0
Flask-CORS==4.0.0

# Database (For future use)
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# Background Task Processing (Celery)
celery==5.3.4
redis==5.0.0

# Testing Framework
pytest==7.4.2
pytest-flask==1.3.0  # Compatible with Flask 3.x
pytest-cov==4.1.0

# Development Tools
black==23.9.1  # Code formatter
flake8==6.1.0  # Linter

# Production Server
gunicorn==21.2.0

# ========================================
# 3D MODELING & MESH PROCESSING
# ========================================

# Image processing for cookie cutters
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.1.0

# 3D mesh processing
trimesh==4.0.5
shapely==2.0.2
mapbox-earcut  # Required by trimesh for polygon triangulation

# Mesh repair and processing
pymeshlab==2023.12.post3
# open3d==0.18.0    # For advanced point cloud/mesh ops (optional)

# 2D geometry operations
pyclipper==1.3.0.post5

# File format support
# ezdxf==1.1.0      # For DXF import/export (optional)
# pygltflib==1.16.1 # For GLTF/GLB export (optional)

# Parametric CAD (will be used for OpenSCAD-like features)
# build123d==0.5.0  # Uncomment when ready to use
# cadquery==2.4.0   # Alternative option

# Math and scientific computing
scipy==1.11.4

# ========================================
# PEOPLE FINDER - CORE DEPENDENCIES
# ========================================
aiohttp==3.9.1           # Async HTTP client
beautifulsoup4==4.12.2   # HTML parsing
lxml==4.9.3              # XML/HTML parser
requests>=2.32.2         # HTTP requests
html5lib==1.1            # HTML5 parser
python-Levenshtein==0.23.0  # Fuzzy string matching for de-duplication
reportlab==4.0.7         # PDF generation for export reports

# Optional performance boost for async (not for Windows)
uvloop==0.19.0; sys_platform != 'win32'

# ========================================
# PEOPLE FINDER - ML/NLP (OPTIONAL)
# Pre-trained models - NO training required!
# ========================================

# Sentence-BERT for semantic name matching
# Model: all-MiniLM-L6-v2 (90MB download on first run)
# Handles: Name similarity, nickname detection, typo handling
sentence-transformers>=2.2.2

# spaCy for Named Entity Recognition (NER)
# Model: en_core_web_lg (large English model - 560MB)
# Handles: Person, date, location, organization extraction from HTML
spacy>=3.7.0

# usaddress for ML-based address parsing
# Pre-trained on millions of US addresses
# Handles: Non-standard address formats, component extraction
usaddress>=0.5.10

# ========================================
# POST-INSTALLATION STEPS
# ========================================

# After installing, download spaCy model:
#   python -m spacy download en_core_web_lg

# Verify ML models work:
#   python -c "from sentence_transformers import SentenceTransformer; print('✓ Sentence-BERT OK')"
#   python -c "import spacy; spacy.load('en_core_web_lg'); print('✓ spaCy OK')"
#   python -c "import usaddress; print('✓ usaddress OK')"

# ========================================
# OPTIONAL: FUTURE CUSTOM MODEL TRAINING
# (Uncomment if you want to train your own models)
# ========================================

# Google Cloud AI Platform
# google-cloud-aiplatform>=1.25.0

# TensorFlow
# tensorflow>=2.13.0

# PyTorch (alternative to TensorFlow)
# torch>=2.0.0

# Transformers (for fine-tuning BERT models)
# transformers>=4.30.0

# ========================================
# TROUBLESHOOTING
# ========================================

# If sentence-transformers install fails:
#   pip install --upgrade pip setuptools wheel
#   pip install sentence-transformers

# If spaCy model download fails:
#   python -m spacy download en_core_web_lg --user

# If usaddress install fails:
#   pip install python-crfsuite==0.9.7
#   pip install usaddress

# ========================================
# SYSTEM REQUIREMENTS
# ========================================

# Python 3.8 or higher
# ~2GB RAM for ML model loading
# ~1.5GB disk space for ML models
# NO GPU required (models run on CPU)
